+++ src/hitable.rs
@@ -5,14 +5,14 @@ use material::Material;
 use std::rc::Rc;
 
 pub struct HitRecord {
-    pub t: f32,
+    pub t: f64,
     pub p: Vec3,
     pub normal: Vec3,
     pub mat: Rc<Material>
 }
 
 impl HitRecord {
-    pub fn new(t: f32, p: Vec3, normal: Vec3, mat: Rc<Material>) -> HitRecord {
+    pub fn new(t: f64, p: Vec3, normal: Vec3, mat: Rc<Material>) -> HitRecord {
         HitRecord{
             t: t,
             p: p,
@@ -23,5 +23,5 @@ impl HitRecord {
 }
 
 pub trait Hitable {
-    fn hit(&self, ray: &Ray, tmin: f32, tmax: f32) -> Option<HitRecord>;
+    fn hit(&self, ray: &Ray, tmin: f64, tmax: f64) -> Option<HitRecord>;
 }
